Start new run when run_id don't match
Consider an engine connected to an aggregator. A run is started. The aggregator goes offline. Meanwhile, a new run is started on the engine. When the aggregator comes back online then it detects the difference in the run ID reported by the engine and the one that was previously persisted. In this case it is appropriate to allocate the new current run on the aggregator.